Add field for referral source activity
