+++ src/apps/investments/client/projects/create/InvestmentProjectForm.jsx
@@ -57,7 +57,7 @@ const InvestmentProjectForm = ({ company, csrfToken }) => (
             }
             flashMessage={() => 'Investment project created'}
           >
-            {({ values, errors }) => {
+            {({ values }) => {
               const dhCompany = company || values.company
               return (
                 <>
@@ -66,11 +66,7 @@ const InvestmentProjectForm = ({ company, csrfToken }) => (
                   <InvestmentTypeStep
                     investmentTypes={values.investmentTypes}
                   />
-                  <InvestmentDetailsStep
-                    company={dhCompany}
-                    values={values}
-                    errors={errors}
-                  />
+                  <InvestmentDetailsStep company={dhCompany} values={values} />
                 </>
               )
             }}
